+++ android/src/main/java/com/twiliovoicereactnative/CallRecordDatabase.java
@@ -17,15 +17,19 @@ import com.twilio.voice.Call;
 import com.twilio.voice.CallException;
 import com.twilio.voice.CallInvite;
 import com.twilio.voice.CancelledCallInvite;
+import com.twilio.voice.IceOptions;
 
 class CallRecordDatabase  {
   public static class CallRecord {
     public enum CallInviteState { NONE, ACTIVE, USED }
     public enum Direction { INCOMING, OUTGOING }
+    // No notification raised
+    // notifications in this library are always assigned a value >= 1.
+    public static final int INVALID_NOTIFICATION_ID = -1;
     private final UUID uuid;
     private String callSid = null;
     private Date timestamp = null;
-    private int notificationId = -1;
+    private int notificationId = INVALID_NOTIFICATION_ID;
     private Call voiceCall = null;
     private String callRecipient = "";
     private CallInvite callInvite = null;
@@ -33,6 +37,7 @@ class CallRecordDatabase  {
     private CancelledCallInvite cancelledCallInvite = null;
     private ModuleProxy.UniversalPromise callAcceptedPromise = null;
     private ModuleProxy.UniversalPromise callRejectedPromise = null;
+    private IceOptions iceOptions = null;
     private CallException callException = null;
     private Map<String, String> customParameters = null;
     private String notificationDisplayName = null;
@@ -108,6 +113,9 @@ class CallRecordDatabase  {
     public ModuleProxy.UniversalPromise getCallRejectedPromise() {
       return this.callRejectedPromise;
     }
+    public IceOptions getIceOptions() {
+      return this.iceOptions;
+    }
     public CallException getCallException() {
       return this.callException;
     }
@@ -140,6 +148,9 @@ class CallRecordDatabase  {
     }
     public void setCallRejectedPromise(@NonNull ModuleProxy.UniversalPromise callRejectedPromise) {
       this.callRejectedPromise = callRejectedPromise;
+    }
+    public void setIceOptions(IceOptions iceOptions) {
+      this.iceOptions = iceOptions;
     }
     public void setCallException(CallException callException) {
       this.callException = callException;